Haystack should not silently fail and return empty search results
Now an internal server error is shown to the user if there should be any
search backend problem. Per Django default, such a problem is
automatically reported to the server administrator.

Previously the user always just received empty search results, which
might be confusing. In addition we have now the possibility to make sure
that Haystack has a working search backend on server startup.