add “default_retriever” mechanism to `DictField`
Every model with a `DictField` can be instructed to return the default
value of the dictionary using the `get_default_FOO()` method (where
`FOO` is the name of the field on the model). By default, the default
value will be the unicode representation of a random value from the
dictionary or the empty string if the dictionary is empty. The mechanism
which determines the default value may be overridden.