do not hardcode the metadata XML Schema namespace and fix #205
The metadata XML Schema namespace is taken from the new constant in
`models.py` instead of hardcoding it in various places of the other
Python code.

In the exporter we now create an `xsi:schemaLocation` with a URL which
is based on the new schema version constant from `models.py` (fixes
issue #205).