make editing editor groups, organizations and their managers more usable
Editor pages for editor groups and editor group managers do not expose the permissions list anymore as this will usually not be required. If any permissions should need to be changed here, then this can be done via the `Group` admin form. This fixes #445.

Editor pages for organizations and organization managers still expose the permissions list as adding, for example, a META-SHARE membership permission to those groups should be a common use case. However, the permissions list is more usable now as it uses a different widget (as suggested in #445).