make sure to remove resources from the index when they are unpublished
Just updating an ingested/unpublished resource only worked partially and led to strange results in the search results view: the resource is not visible anymore and the results count is correct on the results page where the resource would have appeared. However, the results count on other search result pages is not correct.

Completely removing an unpublished resource from the index solves the problem.