when updating a storage object, then also update its source URL
We now always make sure that any changes to the `DJANGO_URL` setting are also reflected in the `source_url` field of master copies. Otherwise, when a node was not correctly configured or moves to another URL, then the `source_url` would keep the old value. That could then lead to redirection problems on nodes that fetch metadata from us.