workaround for #620: compute binary LR data checksum only when necessary
For large binary LR data (e.g. >100GB), the checksum computation may
take a long time. Instead of computing the checksum on each storage
update, we only (re)compute the checksum, when the binary data changes.

This hopefully fixes #620 and related problems for the moment.

In the future we should review the checksum usage in general: is the
checksum field required at all? It appears to not really be used at the
moment.