further reduce the binary LR data checksum computations to a minimum
For large binary LR data (e.g. >100GB), the checksum computation may
take a long time. Computing the checksum each time a download is
requested in the UI will probably not work: the user will have to wait a
very long time for the download to start. What is more, the calculated
checksum is not really used in such cases!

With this commit, the checksums are only computed when binary data is
uploaded or added to the storage folder. However, as the checksums are
not used anywhere, we should review the checksum usage in the future: is
the checksum field required at all?