Update dependencies to recent versions compatible with django-1.3.x
 - six
 - Unidecode
 - django-kronos
 - pycountry
 - requests

Some of the modified dependencies depend on github.com/zeehio/*.
I won't delete the repositories, but I believe it would make sense
for metashare to host their own copy under github.com/metashare
just in case.

Additionally, in the long term the code should be ported to depend on
standard versions of the dependencies.